refactor(i18n): move i18n related files into i18n folder
