+++ src/i18n/DopplerIntlProvider.double-with-ids-as-values.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import { addLocaleData, IntlProvider } from 'react-intl';
 import en from 'react-intl/locale-data/en';
-import messages_en from './i18n/en.json';
-import { flattenMessages } from './utils';
+import messages_en from './en.json';
+import { flattenMessages } from '../utils';
 
 addLocaleData([...en]);
