+++ src/i18n/DopplerIntlProvider.js
@@ -2,9 +2,9 @@ import React from 'react';
 import { addLocaleData, IntlProvider } from 'react-intl';
 import en from 'react-intl/locale-data/en';
 import es from 'react-intl/locale-data/es';
-import messages_es from './i18n/es.json';
-import messages_en from './i18n/en.json';
-import { flattenMessages } from './utils';
+import messages_es from './es.json';
+import messages_en from './en.json';
+import { flattenMessages } from '../utils';
 
 const messages = {
   es: messages_es,
